+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/service/FitnessClassService.java
@@ -2,5 +2,5 @@ package cz.cvut.fit.tjv.fitnessApp.service;
 
 import cz.cvut.fit.tjv.fitnessApp.domain.FitnessClass;
 
-public interface FitnessClassService extends CrudService<FitnessClass, Integer> {
+public interface FitnessClassService extends CrudService<FitnessClass, Long> {
 }
